ml-server/app/requirements.txt
chromadb==0.5.23
Flask==3.1.0
Flask_Cors==5.0.0
langchain_community==0.3.13
langchain_core==0.3.28
langchain_groq==0.2.2
pandas==2.0.2
pymongo==4.10.1
python-dotenv==1.0.1
